% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\name{construct_ellipses}
\alias{construct_ellipses}
\title{Construct standard ellipses from fitted model}
\usage{
construct_ellipses(mfit, newdat, n = 1, q = 1)
}
\arguments{
\item{mfit}{Fitted isoniche model object.}

\item{newdat}{New data, with all the same columns as the original data used to
fit the model.}

\item{n}{Number of draws from the posterior that can be used to display standard ellipses.
The default is 1, which results in ellipses being drawn from the marginal means of the posterior
of each parameter. If \code{n} > 1, then the function will draw \code{n} samples from the joint
posterior and construct a standard ellipse for each set of parameters.}

\item{q}{Quantile to use for ellipses. For standard ellipses, \code{q = 1},
For approximate 95\% ellipses, you could use \code{qchisq(0.95, 2)}.}
}
\value{
A dataframe that can be used for plotting standard ellipses or calculating isotopic niche
statistics.
}
\description{
This function takes in a dataframe that represents the new data over which
to predict. For example, a set of groups for which to construct standard
ellipses.
}
\examples{
#' \dontrun{
library(isoniche)
library(dplyr)
library(ggplot2)
set.seed(9986)

# Load the example data
data("eg_data")

# ggplot(dat, aes(x = V1, y = V2, color = grp)) +
#   geom_point()

# fitting the model, one bivariate normal
# distribution per group
mfit <- isoniche(
  mean = list(
    y1 ~ grp + x,
    y2 ~ grp + x
  ),
  var = list(
    ~ grp,
    ~ grp,
    ~ grp
  ),
  data = eg_data
)

# make new data for constructing ellipses
newdat <- data.frame(
  grp = unique(eg_data$grp),
  x = rep(0, length(unique(eg_data$grp)))
)

plotdf <- construct_ellipses(mfit, newdat, n = 50, q = 6)

ggplot(data = eg_data, aes(x = y1, y = y2, color = grp)) +
  geom_point() +
  geom_path(
    data = plotdf,
    aes(group = ellipse_id),
    size = 0.2
  ) +
  theme_classic() +
  scale_color_manual(values = c("steelblue", "brown"))

sea_post <- sea(mfit, newdat, n = 250)

ggplot(sea_post, aes(x = sea)) +
  geom_density(aes(fill = grp), alpha = 0.5, color = "grey") +
  theme_classic() +
  scale_fill_manual(values = c("steelblue", "brown"))
#' }
}
