% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\name{ellipse_overlap_mc}
\alias{ellipse_overlap_mc}
\title{Estimate Posterior Overlap of Pairs of Ellipses}
\usage{
ellipse_overlap_mc(mfit, df_pairs, q = 1, n = 250, npp = 1000)
}
\arguments{
\item{mfit}{An \code{isoniche} model object.}

\item{df_pairs}{A data frame or a list of data frames. If a list of dataframes,
each should have 2 rows defining the comparisons. Each row represents a condition
for which the ellipse is defined. If a single data frame, the function will compute all pairwise
overlaps by default.}

\item{q}{The quantile used to define the ellipses. This can be defined based on
the desired quantile of a \eqn{\chi^2_2} distribution. The default is 1.}

\item{n}{An integer specifying the number of draws from the posterior distribution. Default is 250.}

\item{npp}{An integer specifying the number of points to be used in the Monte Carlo simulation for
estimating ellipse overlap. Default is 1000. Overlap estimates will be more precise with more draws,
but the function will get slow.}
}
\value{
A list where each element corresponds to a pair of conditions and contains the following components:
\itemize{
\item \code{cond1}: The first condition in the pair.
\item \code{cond2}: The second condition in the pair.
\item \code{overlap_area}: A numeric vector of estimated overlap areas for each draw from the posterior.
This approximates the posterior distribution of the area of overlap.
}
}
\description{
This function estimates the overlap area for each pair of ellipses
defined by the user for each of \code{n} draws from the posterior distribution
of the mean vector and covariance matrix defining each ellipse.
}
\details{
For pair of conditions and each draw from the posterior distribution, the function first
simulates \code{npp} points uniformly within a bounding box that contains the ellipses.
It then tests whether these points lie within each of the ellipses defined
by their respective means and covariance matrices. The overlap area is calculated based
on the proportion of points that lie within all ellipses, multiplied by the area of the
bounding box.
}
\seealso{
\code{\link{ellipse_overlap_single_mc}}
}
