% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\name{ellipse_overlap_single_mc}
\alias{ellipse_overlap_single_mc}
\title{Calculate the overlap of ellipses using Monte Carlo simulation}
\usage{
ellipse_overlap_single_mc(mus, Sigmas, q, npp = 1000)
}
\arguments{
\item{mus}{A list of numeric vectors representing the mean vectors of the ellipses.}

\item{Sigmas}{A list of matrices representing the covariance matrices of the ellipses.}

\item{q}{A numeric value representing the quantile used to define the ellipses.
This can be defined based on the desired quantile of a \eqn{\chi^2_2} distribution.}

\item{npp}{An integer specifying the number of points to be used in the Monte Carlo simulation.
Default is 1000.}
}
\value{
A numeric value representing the estimated overlap area of the ellipses.
}
\description{
This function estimates the overlap area of multiple ellipses
defined by their means and covariance matrices using a Monte Carlo
simulation approach.
}
\details{
The function first simulates \code{npp} points uniformly within a bounding box that contains
the ellipses. It then tests whether these points lie within each of the ellipses defined
by their respective means and covariance matrices. The overlap area is calculated based
on the proportion of points that lie within all ellipses, multiplied by the area of the
bounding box.
}
\examples{
\dontrun{
mus <- list(c(0, 0), c(1, 1))
Sigmas <- list(diag(2), diag(2))
q <- 2
ellipse_overlap_single_mc(mus, Sigmas, q)
}

}
