% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iai.R
\name{iai}
\alias{iai}
\title{Compute the isodar-adjusted inequality (IAI) index}
\usage{
iai(
  isodars,
  abundances = NULL,
  weights = NULL,
  method = "gini",
  plot = FALSE,
  scaled = TRUE,
  max_search = 10000,
  dim = "pairwise"
)
}
\arguments{
\item{isodars}{A data frame returned by fit_isodars().}

\item{abundances}{Numeric vector of total abundances. If NULL, the minimum
total abundance yielding positive abundance in all habitats is searched.}

\item{weights}{Weighting scheme for pairwise reconstruction. Either NULL,
a numeric vector, or "1/var". Not used for dim = "ndim" (1 isodar - no variance).}

\item{method}{Currently only "gini".}

\item{plot}{Logical. If TRUE, plots niche breadth against chosen total abundances.}

\item{scaled}{Logical. If TRUE, coerces minimum possible IAI to be zero rather than 1/n.}

\item{max_search}{Maximum abundance searched when abundances = NULL.}

\item{dim}{Either "pairwise" or "ndim". Should match the dim used in fit_isodars().}
}
\value{
A tibble containing total abundance, niche breadth, and the
reconstructed abundance of each habitat.
}
\description{
Reconstructs habitat abundances for one or more requested total
abundances using either pairwise or n-dimensional  isodars.
}
\details{
With dim = "pairwise", the pairwise relationships are simultaneously
fitted using (optionally weighted) total least squares (TLS).

With dim = "ndim", the n-dimensional TLS hyperplane is used directly.
For a given total abundance, the reconstructed abundance vector is the
point on the TLS hyperplane and total-abundance plane that is closest to
equal allocation among habitats. Non-negative abundances are enforced
using an active-set procedure.

Niche breadth is defined as 1 - Gini(x).
}
\examples{
set.seed(1)
populations1 <- simulate_isodars(1, 1.1, 1, 0, noise = 2, n = 10)
 isodars1 <- fit_isodars(populations1, hab_cols=c("hab1","hab2","hab3"))
  iai(isodars1)
   iai(isodars1,abundances=c(6,10,16,20,40,60,130,260),weights="1/var",plot=TRUE)
populations2 <- simulate_isodars(1.5, 2, 5, 2, noise = 2, n = 10)
 isodars2 <- fit_isodars(populations2, hab_cols=c("hab1","hab2","hab3"), dim="ndim")
  iai(isodars2, dim="ndim", scaled = "FALSE")
   iai(isodars2, dim="ndim")
    iai(isodars2, dim="ndim",abundances=c(26,46,60,106,160,260),plot=TRUE)

}
