% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodar_adj_niche.R
\name{isodar_adj_niche}
\alias{isodar_adj_niche}
\title{Compute the isodar-adjusted inequality index for one or more total abundances}
\usage{
isodar_adj_niche(
  isodars,
  abundances = NULL,
  weights = NULL,
  alpha = 0.1,
  sig_weights = c(sig = 1, nonsig = 0),
  method = c("gini"),
  plot = FALSE,
  max_search = 10000
)
}
\arguments{
\item{isodars}{A data frame returned by \code{fit_isodar()}}

\item{abundances}{Numeric vector of total abundances. If NULL, the minimal total
abundance yielding occupancy in all habitats ("baseline") is chosen when possible; otherwise
a default exploratory sequence is used}

\item{weights}{Weighting scheme. One of:
\itemize{
\item \code{NULL}: equal weights
\item numeric vector of length \code{nrow(isodars)}
\item \code{"1/var"}: weights proportional to \code{1/sd^2}
\item \code{"sig"}: weights based on significance of intercepts, using \code{sig_weights}
}}

\item{alpha}{Numeric in (0, 1). Significance threshold used when \code{weights = "sig"}}

\item{sig_weights}{Numeric vector of length 2 giving weights for significant and non-significant
isodars when \code{weights = "sig"}. Can be named \code{c(sig = , nonsig = )} or unnamed
\code{c(sig, nonsig)}. Default is \code{c(sig = 1, nonsig = 0)}.}

\item{method}{Character. Currently only \code{method = "gini"} available}

\item{plot}{Logical. If TRUE, plots adjusted niche breadth versus total abundance}

\item{max_search}{Integer. Maximum total abundance to search when \code{abundances = NULL}, in order to find baseline n}
}
\value{
A tibble with one row per total abundance containing:
\itemize{
\item \code{total_abundance}
\item niche_breadth
\item one column per habitat with reconstructed abundance
}
}
\description{
Reconstructs habitat abundances for each requested total abundance by solving a
weighted least-squares system defined by the pairwise isodars:
\deqn{habitat\_y = intercept + slope \times habitat\_x }
}
\details{
The reconstruction is constrained to be nonnegative and to sum to the requested
total abundance (active-set heuristic). Niche breadth is then computed
from the reconstructed habitat vector; currently defined as:
\deqn{niche breadth = 1 - Gini(x)} of isodar-reconstructed abundances at a given N.

To emphasize the intercept component, use baseline abundances provided by \code{abundances = NULL}.
To emphasize the slope component, we recommend to set \code{abundances} to the highest n in your dataset.
To see how niche breadth changes as a function of n, enter a vector of increasing abundances.
}
\examples{
set.seed(1)
isod <- simulate_isodars(1, 2, 5, 1, noise = 2, n = 10)
INB <- fit_isodar(isod, n_habitats = 3)
isodar_adj_niche(INB, max_search = 100) # automatically checks only baseline abundance
isodar_adj_niche(INB, abundances = c(15, 30, 45), max_search = 100) # set specific values

}
