% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.isoniche}
\alias{plot.isoniche}
\title{Plot posterior means of residual distributions}
\usage{
\method{plot}{isoniche}(mfit)
}
\arguments{
\item{mfit}{Fitted isoniche object}
}
\value{
Multipanel plot of residuals. The first shows the bivariate residuals, which should
be approximately spherical (i.e., be bivariate normal with mean zero and covariance matrix
\eqn{I_n}). The next plots show qq plots for each dimension.
}
\description{
Plot posterior means of residual distributions
}
