% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.isoniche}
\alias{print.summary.isoniche}
\title{Print methods for the summarized isoniche object}
\usage{
\method{print}{summary.isoniche}(sumobj, digits = 3)
}
\arguments{
\item{sumobj}{summary.isoniche object}

\item{digits}{Number of digits to include in printed table. Defaults to 3.}
}
\description{
Print methods for the summarized isoniche object
}
