% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.isoniche}
\alias{residuals.isoniche}
\title{Mean of posterior normalized residual distribution}
\usage{
\method{residuals}{isoniche}(mfit)
}
\arguments{
\item{mfit}{Fitted isoniche object}
}
\value{
Matrix of posterior mean residuals, one column for each response
variable.
}
\description{
Mean of posterior normalized residual distribution
}
