% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\name{sea}
\alias{sea}
\title{Construct posterior of SEA}
\usage{
sea(mfit, newdat, n = 250)
}
\arguments{
\item{mfit}{Fitted isoniche model object.}

\item{newdat}{New data, with all the same columns as the original data used to
fit the model.}

\item{n}{Number of draws from the posterior SEA for each group or condition in \code{newdat}.}
}
\value{
A dataframe that can be used for plotting posterior SEAs or computing summary statistics.
}
\description{
This function takes in a dataframe that represents the groups or conditions over which to compute
the standard ellipse area (SEA) and takes \code{n} draws from the posterior covariance matrix
for each row of \code{newdat}. It then computes the SEA for each and returns a dataframe
of the posterior draws with one row per draw per condition.
}
\examples{
#' \dontrun{
library(isoniche)
library(dplyr)
library(ggplot2)
set.seed(9986)

# Load the example data
data("eg_data")

# ggplot(dat, aes(x = V1, y = V2, color = grp)) +
#   geom_point()

# fitting the model, one bivariate normal
# distribution per group
mfit <- isoniche(
  mean = list(
    y1 ~ grp + x,
    y2 ~ grp + x
  ),
  var = list(
    ~ grp,
    ~ grp,
    ~ grp
  ),
  data = eg_data
)

# make new data for constructing ellipses
newdat <- data.frame(
  grp = unique(eg_data$grp),
  x = rep(0, length(unique(eg_data$grp)))
)

plotdf <- construct_ellipses(mfit, newdat, n = 50, q = 6)

ggplot(data = eg_data, aes(x = y1, y = y2, color = grp)) +
  geom_point() +
  geom_path(
    data = plotdf,
    aes(group = ellipse_id),
    size = 0.2
  ) +
  theme_classic() +
  scale_color_manual(values = c("steelblue", "brown"))

sea_post <- sea(mfit, newdat, n = 250)

ggplot(sea_post, aes(x = sea)) +
  geom_density(aes(fill = grp), alpha = 0.5, color = "grey") +
  theme_classic() +
  scale_fill_manual(values = c("steelblue", "brown"))
#' }
}
