% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isodars.R
\name{simulate_isodars}
\alias{simulate_isodars}
\title{Simulate abundances across three habitats}
\usage{
simulate_isodars(
  slope1,
  slope2,
  int1,
  int2,
  noise = 1,
  n = 30,
  hab1_max = 100/(slope1^1.2) - int1
)
}
\arguments{
\item{slope1}{Numeric. Isodar slope for \code{hab2 ~ hab1}.}

\item{slope2}{Numeric. Isodar slope for \code{hab3 ~ hab2}.}

\item{int1}{Numeric. Isodar intercept for \code{hab2 ~ hab1}.}

\item{int2}{Numeric. Isodar intercept for \code{hab3 ~ hab2}.}

\item{noise}{Numeric (>= 0). Standard deviation of Gaussian noise.}

\item{n}{Integer (> 0). Number of simulated sites.}

\item{hab1_max}{Integer. maximum abundance value in habitat 1 (other maxima calculated from hab1_max).
Default is derived to generally suit the isodar parameters, but is arbitrary.}
}
\value{
A tibble with one row per simulated site containing:
\itemize{
\item hab1, hab2, hab3: simulated abundances (nonnegative integers)
\item total_abundance: hab1 + hab2 + hab3
\item mean_abundance: total_abundance divided  by 3
\item sd: scaled standard deviation of habitat abundances
\item cv: coefficient of variation (NA if mean_abundance == 0)
\item gini: Gini coefficient of habitat abundances
}
}
\description{
Simulates abundance data for three habitats across \code{n} sites using
sequential linear relationships (isodars) with Gaussian noise:
\itemize{
\item hab2 is generated from hab1 using \code{slope1} and \code{int1}
\item hab3 is generated from hab2 using \code{slope2} and \code{int2}
}
}
\details{
The output includes simple per-site niche metrics (CV and Gini) computed
from the three habitat abundances.
}
\examples{
set.seed(1)
simulate_isodars(1, 2, 5, 1, noise = 2, n = 10)

}
