% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.isoniche}
\alias{summary.isoniche}
\title{Summarize fitted isoniche model}
\usage{
\method{summary}{isoniche}(mfit)
}
\arguments{
\item{mfit}{A fitted isoniche object}
}
\value{
This general method returns a summary of the results of the fitted model, including
95\% credible intervals for model coefficients as well as Gelman's Rhat statistic for
estimating chain convergence.
}
\description{
Summarize fitted isoniche model
}
