% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\docType{package}
\name{isoniche}
\alias{isoniche-package}
\alias{isoniche}
\title{isoniche: Calculating Density-Independent Niche Breadth Indices from Abundance Data}
\description{
Deriving isodar-based niche breadth indices from
abundance data of two or more habitats, including several methods based on
pairwise isodars, multidimensional isodars, and isodar-adjusted inequality.
}
\author{
\strong{Maintainer}: Shahar Dubiner \email{dubiner@mail.tau.ac.il} (\href{https://orcid.org/0000-0003-1771-0087}{ORCID})

Authors:
\itemize{
  \item Itai Granot
  \item Jonathan Belmaker
}

}
