% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoniche.R
\name{isoniche}
\alias{isoniche}
\title{isoniche}
\usage{
isoniche(mean, var, data, ...)
}
\arguments{
\item{mean}{A list of length 2 with formulas defining the models for the means along
each axis.}

\item{var}{A list of length 3 of one-sided formulas that define the models for the
scale parameters (standard deviations) along each axis as well as a model for the correlation
parameter. These must be defined, but can be set to \code{~ 1} for constant scales or correlation
across groups / individuals.}

\item{data}{Data frame with one row per individual and columns for the isotope concentrations as
well as group designations (as factors) and any control variables.}

\item{...}{Additional arguments that can be passed to the sampler. These include \code{iter}, \code{chains},
\code{thin}, \code{cores}, \code{control}. See \code{?rstan::sampling()} for more details.}
}
\value{
A fitted model object from the \pkg{rstan} package.
}
\description{
Function to fit a bivariate regression model for calculating standard ellipses
in isotopic niche comparisons.
}
